Alerts redefined

+++ public/js/main/components/Fragments/AlertDelete.jsx
@@ -1,46 +1,26 @@
-import * as React from "react";
-import Button from "@mui/material/Button";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogContentText from "@mui/material/DialogContentText";
-import DialogTitle from "@mui/material/DialogTitle";
-import Slide from "@mui/material/Slide";
+import React from "react";
+import Swal from "sweetalert2";
 
-const Transition = React.forwardRef(function Transition(props, ref) {
-  return <Slide direction="up" ref={ref} {...props} />;
-});
+export default function AlertDelete({ handleDelete }) {
+  const handleConfirmDelete = () => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        handleDelete(); // Call the delete function if the user confirms
+      }
+    });
+  };
 
-export default function AlertDelete({ handleDelete, handleClickOpen, open, handleClose }) {
   return (
     <React.Fragment>
-      <Dialog
-        open={open}
-        TransitionComponent={Transition}
-        keepMounted
-        onClose={handleClose}
-        aria-describedby="alert-dialog-slide-description"
-      >
-        <DialogTitle>{"Sure to delete this record?"}</DialogTitle>
-        <DialogContent>
-          <DialogContentText id="alert-dialog-slide-description">
-            {" "}
-            This action cannot be undone. Warning: Deleting this item will permanently remove it from the system. Please
-            ensure that you have backed up any important data before proceeding.
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose}>Keep</Button>
-          <Button
-            onClick={() => {
-              handleDelete();
-              handleClose();
-            }}
-          >
-            Wipe Data
-          </Button>
-        </DialogActions>
-      </Dialog>
+      {/* The SweetAlert2 dialog will be triggered by handleConfirmDelete, so no need for a Dialog component here. */}
     </React.Fragment>
   );
 }
